+++ api/views/spark.py
@@ -234,6 +234,25 @@ def delete(request, device_id):
 
     return HttpResponse('{"Status":"OK"}\n', content_type="application/json")
 
+@require_http_methods(["DELETE"])
+def device_delete(request, device_id, actuator_id):
+    logger.info("Delete device {} from spark {} from database".format(actuator_id, device_id))
+
+    try:
+        device = Device.objects.get(pk=actuator_id)
+    except ObjectDoesNotExist:
+        return HttpResponse('{"Status":"ERROR",Message="Device does not exists"}\n',
+                            content_type="application/json",
+                            status=400)
+
+    if device.spark.device_id != device_id:
+        return HttpResponse('{"Status":"ERROR",Message="Device not assigned to given Spark"}\n',
+                            content_type="application/json",
+                            status=400)
+
+    device.delete()
+
+    return HttpResponse('{"Status":"OK"}\n', content_type="application/json")
 
 @require_http_methods(["POST"])
 def device_toggle(request, device_id, actuator_id):